Route parameters should overwrite all other merged variables:
When merging variables, body parameters overwrite query paraemeters,
but they would both overwrite the route parameters.

This commit makes sure (and adds a test to make sure) that route
parameters will not be overwritten.

This addresses GH #931, except it doesn't allow the query
parameters to overwrite the body.

Seems like Express.js does Query > Body, and Rails does Body > Query
and it's probably - if anything - not RFC compliant to begin with.